Child's Potty Dataset
Dataset sourcing from my child's daycare, I further aggregated the data into an increment of every 30 minutes. I plotted the poop in column chart combining with the pee counts in line-chart. 
This resulting in a conclusion of my child's highest frequency in pooping around 8:30-10:30 am, 12:00 - 12:30 pm, and 3:30-4:00 pm. 
This schedule can be best used to direct staff in helping my child to go to potty at the appropriate time.